--- pkg/query-service/dao/interfaces/interface.go
@@ -1,6 +1,0 @@
-package interfaces
-
-type ModelDao interface {
-	Queries
-	Mutations
-}
